+++ src/lib/components/table/transactions/TransactionsTableMobileCard.tsx
@@ -51,8 +51,13 @@ export const TransactionsTableMobileCard = ({
         <Flex>
           <RenderActionMessages transaction={transaction} />
           {transaction.isIbc && (
-            <Tag variant="accent-dark" size="sm" w={8} justifyContent="center">
+            <Tag variant="accent-dark" size="sm">
               IBC
+            </Tag>
+          )}
+          {transaction.isOpinit && (
+            <Tag variant="teal" size="sm">
+              OPInit
             </Tag>
           )}
         </Flex>