+++ src/lib/pages/past-txs/index.tsx
@@ -21,10 +21,9 @@ import { TransactionsTableWithWallet } from "lib/components/table";
 import { TxFilterSelection } from "lib/components/TxFilterSelection";
 import { TxRelationSelection } from "lib/components/TxRelationSelection";
 import { DEFAULT_TX_FILTERS } from "lib/data";
-import { useAccountId } from "lib/services/accountService";
 import {
-  useTxsByAddressPagination,
-  useTxsCountByAddress,
+  useAPITxsCountByAddress,
+  useTxsByAddress,
 } from "lib/services/txService";
 import type { HumanAddr, Option, TxFilters } from "lib/types";
 
@@ -52,16 +51,15 @@ const PastTxs = () => {
     defaultValues,
     mode: "all",
   });
-
   const pastTxsState = watch();
 
-  const { data: accountId } = useAccountId(address as HumanAddr);
-  const { data: countTxs = 0 } = useTxsCountByAddress({
-    accountId,
-    search: pastTxsState.search,
-    filters: pastTxsState.filters,
-    isSigner: pastTxsState.isSigner,
-  });
+  const { data: rawTxCount } = useAPITxsCountByAddress(
+    address as HumanAddr,
+    pastTxsState.search,
+    pastTxsState.isSigner,
+    pastTxsState.filters
+  );
+  const txCount = rawTxCount ?? undefined;
 
   const {
     pagesQuantity,
@@ -71,7 +69,7 @@ const PastTxs = () => {
     setPageSize,
     offset,
   } = usePaginator({
-    total: countTxs,
+    total: txCount,
     initialState: {
       pageSize: 10,
       currentPage: 1,
@@ -79,62 +77,44 @@ const PastTxs = () => {
     },
   });
 
-  const { data: txs, isLoading } = useTxsByAddressPagination(
-    accountId,
+  const { data: txs, isLoading } = useTxsByAddress(
+    address as HumanAddr,
     pastTxsState.search,
-    pastTxsState.filters,
     pastTxsState.isSigner,
+    pastTxsState.filters,
     offset,
     pageSize
   );
 
-  const resetPagination = () => {
-    setPageSize(10);
+  const handleOnSearchChange = (e: ChangeEvent<HTMLInputElement>) => {
     setCurrentPage(1);
+    setValue("search", e.target.value);
   };
 
-  const setFilter = (filter: string, bool: boolean) => {
-    resetPagination();
+  const handleOnIsSignerChange = (value: Option<boolean>) => {
+    setCurrentPage(1);
+    setValue("isSigner", value);
+  };
+
+  const handleOnFiltersChange = (filter: string, bool: boolean) => {
+    setCurrentPage(1);
     setValue("filters", { ...pastTxsState.filters, [filter]: bool });
   };
 
-  const onPageChange = (nextPage: number) => {
-    setCurrentPage(nextPage);
-  };
-
-  const onPageSizeChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    const size = Number(e.target.value);
-    setPageSize(size);
-    setCurrentPage(1);
-  };
-
-  const filterSelected = useMemo(() => {
-    return Object.keys(pastTxsState.filters).reduce(
-      (acc: string[], key: string) => {
+  const filterSelected = useMemo(
+    () =>
+      Object.keys(pastTxsState.filters).reduce((acc: string[], key: string) => {
         if (pastTxsState.filters[key as keyof typeof pastTxsState.filters]) {
           acc.push(key);
         }
         return acc;
-      },
-      []
-    );
-  }, [pastTxsState]);
+      }, []),
+    [pastTxsState]
+  );
 
   useEffect(() => {
     if (router.isReady) track(AmpEvent.TO_PAST_TXS);
   }, [router.isReady]);
-
-  useEffect(() => {
-    setPageSize(10);
-    setCurrentPage(1);
-  }, [
-    chainId,
-    setCurrentPage,
-    setPageSize,
-    pastTxsState.search,
-    pastTxsState.isSigner,
-    pastTxsState.filters,
-  ]);
 
   useEffect(() => {
     reset();
@@ -155,10 +135,7 @@ const PastTxs = () => {
         <InputGroup>
           <Input
             value={pastTxsState.search}
-            onChange={(e) => {
-              setCurrentPage(1);
-              setValue("search", e.target.value);
-            }}
+            onChange={handleOnSearchChange}
             placeholder={`Search with Transaction Hash${
               wasm.enabled ? " or Contract Address" : ""
             }`}
@@ -171,22 +148,19 @@ const PastTxs = () => {
         <Flex gap={3}>
           <TxRelationSelection
             value={pastTxsState.isSigner}
-            setValue={(value) => {
-              resetPagination();
-              setValue("isSigner", value);
-            }}
+            setValue={handleOnIsSignerChange}
             w="165px"
           />
           <TxFilterSelection
             result={filterSelected}
-            setResult={setFilter}
+            setResult={handleOnFiltersChange}
             boxWidth="285px"
             placeholder="All"
           />
         </Flex>
       </Flex>
       <TransactionsTableWithWallet
-        transactions={txs}
+        transactions={txs?.items}
         isLoading={isLoading}
         emptyState={
           pastTxsState.search.trim().length > 0 ||
@@ -194,7 +168,9 @@ const PastTxs = () => {
           filterSelected.length > 0 ? (
             <EmptyState
               imageVariant="not-found"
-              message="No past transaction matches found with your input. You can search with transaction hash, and contract address."
+              message={`No past transaction matches found with your input. You can search with transaction hash${
+                wasm.enabled ? ", and contract address" : ""
+              }.`}
               withBorder
             />
           ) : (
@@ -206,15 +182,19 @@ const PastTxs = () => {
           )
         }
       />
-      {countTxs > 10 && (
+      {!!txCount && txCount > 10 && (
         <Pagination
           currentPage={currentPage}
           pagesQuantity={pagesQuantity}
           offset={offset}
-          totalData={countTxs}
+          totalData={txCount}
           pageSize={pageSize}
-          onPageChange={onPageChange}
-          onPageSizeChange={onPageSizeChange}
+          onPageChange={setCurrentPage}
+          onPageSizeChange={(e) => {
+            const size = Number(e.target.value);
+            setPageSize(size);
+            setCurrentPage(1);
+          }}
         />
       )}
     </PageContainer>